Add rollout strategy to kubernetes config
maxSurge: 100% means that 100% of the number of new replicas is allowed
to be created over the total number of replicas set.

In other words despite the config saying only 2 replicas can exist at
at any one time, during rollout a further 2 are allowed to exist at the
same time.

maxUnavailable: 50% means that at least 50% of the new replicas must be
available before scaling down of the old ones commences