Remove Makefile
Since we are no longer installing the development and test dependencies
in the Dockerfile rspec is not available to run the tests, hence the
change to just run it in the CircleCI config.

In which case the Makefile is longer required. We
also do all the building of images and pushing to ECR via the fb-deploy
scripts.